TASK4
Developing a weather forecast application using Python. This project will involve
retrieving weather data from an API and presenting it in a user-friendly format. The
application will provide current weather information, including temperature, humidity,
wind speed, and a short description of the weather conditions.
User Input: Prompt the user to enter the name of a city or a zip code.
API Request: Make an API request to retrieve weather data based on the user's input.
Parse and Extract: Extract relevant weather information such as temperature, humidity,
and description from the API response.
Display Weather: Present the retrieved weather information to the user in a user-friendly
format